Load full staff info when selected
Introduces functions that pull data in from ajax in json format, then
builds the html from an underscore.js template and inserts it into the
DOM. This base JS also handles all of the html class changes that should
be necessary to handle changing which user is displayed via CSS when a
new user is selected. Functionality for arrowing-through different staff
members is not yet in place but should use the load user/show user
functions.