Implement Default Homepage Content
Adds the default homepage design for users that haven't set up a custom
homepage. This includes:
- News posts as previously designed and implemented at the top for the
first band
- Collections of posts in each featured tutorial tag, with the tutorial
tag title as a heading
- The featured tutorial tags can be set in the Customizer, and can be
changed or reordered there with live previewing at any time
- The Customizer UI for setting these includes a suggestion script that
offers suggested tags as you type, and is housed within a featured
content section that's contextual to the front page
- Tutorials are displayed similarly to posts, but without the author and
date meta, and with the excerpt consisting of the first paragraph of the
content area rather than the post excerpt
- All of the necessary HTML is now being output, so this is ready for
CSS to bring it inline with the designs without further changes to the
underlying PHP/HTML

Next step is to implement the ability to load more items in a given
collection on-the-fly. Then, we can start implementing a screen where
users can create an account and select which items they want on their
personalized home page. At that point we'll need to support terms in
multiple taxonomies here, which will be as simple as making the taxonomy
variable-based instead of hardcoded. I've set this up so that the same
code will be used regardless of whether the default or personalized
bands are shown.